backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh
agno==2.3.21
    # via agents-os (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
backoff==2.2.1
    # via langfuse
beautifulsoup4==4.14.3
    # via wikipedia
brotli==1.1.0
    # via httpx
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   ddgs
    #   rich-toolkit
    #   typer
    #   typer-slim
    #   uvicorn
ddgs==9.6.0
    # via agents-os (pyproject.toml)
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
docstring-parser==0.17.0
    # via agno
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.117.1
    # via agents-os (pyproject.toml)
fastapi-cli==0.0.13
    # via fastapi
fastapi-cloud-cli==0.2.1
    # via fastapi-cli
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via agno
googleapis-common-protos==1.72.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.76.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   agno
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   agno
    #   ddgs
    #   fastapi
    #   fastapi-cloud-cli
    #   langfuse
    #   ollama
    #   openai
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
jinja2==3.1.6
    # via fastapi
jiter==0.11.0
    # via openai
langfuse==3.11.2
    # via agents-os (pyproject.toml)
lxml==6.0.2
    # via ddgs
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==2.3.3
    # via pgvector
ollama==0.6.1
    # via agents-os (pyproject.toml)
openai==1.109.1
    # via
    #   agents-os (pyproject.toml)
    #   langfuse
openinference-instrumentation==0.1.42
    # via openinference-instrumentation-agno
openinference-instrumentation-agno==0.1.25
    # via agents-os (pyproject.toml)
openinference-semantic-conventions==0.1.25
    # via
    #   openinference-instrumentation
    #   openinference-instrumentation-agno
opentelemetry-api==1.39.1
    # via
    #   langfuse
    #   openinference-instrumentation
    #   openinference-instrumentation-agno
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.39.1
    # via agents-os (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.39.1
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.39.1
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.60b1
    # via openinference-instrumentation-agno
opentelemetry-proto==1.39.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.39.1
    # via
    #   agents-os (pyproject.toml)
    #   langfuse
    #   openinference-instrumentation
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.60b1
    # via
    #   openinference-instrumentation-agno
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via
    #   agno
    #   langfuse
    #   opentelemetry-instrumentation
pgvector==0.4.1
    # via agents-os (pyproject.toml)
primp==0.15.0
    # via ddgs
protobuf==6.33.2
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
psycopg==3.2.10
    # via agents-os (pyproject.toml)
psycopg-binary==3.2.10
    # via psycopg
pydantic==2.11.9
    # via
    #   agno
    #   fastapi
    #   fastapi-cloud-cli
    #   langfuse
    #   ollama
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via agno
pygments==2.19.2
    # via rich
python-dotenv==1.1.1
    # via
    #   agno
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via
    #   agno
    #   fastapi
pyyaml==6.0.3
    # via
    #   agno
    #   uvicorn
requests==2.32.5
    # via
    #   agents-os (pyproject.toml)
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-http
    #   wikipedia
rich==14.1.0
    # via
    #   agno
    #   rich-toolkit
    #   typer
    #   typer-slim
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.39.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via
    #   typer
    #   typer-slim
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
socksio==1.0.0
    # via httpx
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.43
    # via agents-os (pyproject.toml)
starlette==0.48.0
    # via fastapi
tqdm==4.67.1
    # via openai
typer==0.21.0
    # via
    #   agno
    #   fastapi-cli
    #   fastapi-cloud-cli
typer-slim==0.21.0
    # via agents-os (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   agno
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   grpcio
    #   openai
    #   openinference-instrumentation-agno
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.37.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wikipedia==1.4.0
    # via agents-os (pyproject.toml)
wrapt==1.17.3
    # via
    #   langfuse
    #   openinference-instrumentation
    #   openinference-instrumentation-agno
    #   opentelemetry-instrumentation
zipp==3.23.0
    # via importlib-metadata
